pyproject.toml
[project]
name = "python-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.18",
    "aiosqlite>=0.21.0",
    "autogen-agentchat>=0.5.6",
    "autogen-ext[azure,magentic-one,ollama,openai]>=0.5.6",
    "autogenstudio>=0.0.49",
    "black>=25.1.0",
    "google-generativeai>=0.8.5",
    "google-search-results>=2.4.2",
    "langchain-community>=0.3.24",
    "langchain-google-genai>=2.0.10",
    "langchain-mcp-adapters>=0.1.7",
    "langchain-openai>=0.3.21",
    "langgraph>=0.4.8",
    "litellm[proxy]>=1.9.4",
    "matplotlib>=3.10.3",
    "mcp>=1.6.0",
    "nest-asyncio>=1.6.0",
    "openai>=1.82.0",
    "pyngrok>=7.2.5",
    "redis==4.6.0",
    "requests>=2.32.3",
    "requests-html>=0.10.0",
    "rich>=14.0.0",
    "uvicorn>=0.34.2",
]
